jax branch fully working
