+++ requirements_dev.txt
@@ -1,3 +1,5 @@
 tox==3.25.1
 pytest==7.1.2
 pytest-cov==3.0.0
+jax>=0.2.7
+jaxlib>=0.1.57
