+++ setup.cfg
@@ -36,6 +36,8 @@ testing =
     pytest>=6.0
     pytest-cov>=2.0
     tox>=3.24
+    jax>=0.2.7
+    jaxlib>=0.1.57
 
 [options.packages.find]
 where = src
