+++ tox.ini
@@ -14,6 +14,10 @@ python =
 setenv =
     PYTHONPATH = {toxinidir}
 deps =
-    -r{toxinidir}/requirements_dev.txt
+    tox==3.25.1
+    pytest==7.1.2
+    pytest-cov==3.0.0
+    jax>=0.2.7
+    jaxlib>=0.1.57
 commands =
     pytest --basetemp={envtmpdir}
